♻️  zb: Split issues-related tests into a separate mod
